Refactored both field to share the same code
Got tired of fixing bugs twice, so this commit creates an abstract class
that contains all the code and contains a basic abstract method in order
to let child classes defined the field name.

Both fields are 100% identical regarding behavior so it did not make
sense to tolerate this longer.