Unify Styling of readonly values
This PR adds the class „field-value-readonly“ to the fields readonly
values in the publish entry area.

This class has been introduced in Symphony 2.6.5 to unify the styling
of readonly values among different extensions.

To avoid raising the minimum requirements for this extension just
because of a styling update I additionally added fallback-styles in the
css. These will be redundant once this extension stops supporting
Symphony 2.6.4 (and therefore can be deleted as soon as this will be
the case).